T0.2.35 1、修改EEPROM存储蓝牙名称出现失败的情况，未完全修复 2、修复 ADC 测量电池电压不准的情况

+++ LightDetect_I2C/LightDetect_I2C.h
@@ -7,7 +7,7 @@ public:
   LIGHTDETECT_I2C(int slave_address);
 
   byte Set_Operate_Mode(byte optMode);
-  byte Get_Light_Value(unsigned short *readValue);
+  byte Get_Light_Value(float *readValue);
 
 private:
   byte _device_address;
