+++ LightDetect_I2C/example/LightDetect_I2C/LightDetect_I2C.ino
@@ -24,7 +24,7 @@ void setup() {
 }
 
 byte lightIndex = 0;
-uint16_t lightValue;
+float lightValue;
 byte errorCode;
 byte ledIndex = 0;
 byte statusValue;
@@ -45,7 +45,7 @@ void loop() {
   errorCode = Light_Sensor.Get_Light_Value(&lightValue);
   if(errorCode != 0){
     Serial.println("### light read Error #################################");
-  }else Serial.printf("*** light Value: %d \n", lightValue);
+  }else Serial.printf("*** light Value: %.2f \n", lightValue);
 
   lightIndex++;
   if(lightIndex > 1) lightIndex = 0;
